refactor gamescreen into board and ui
