+++ src/components/screens/GameScreen.tsx
@@ -1,29 +1,12 @@
-import { useRef } from 'react';
-import { Button, VSpacer, SpriteSheet, GameLoopContextProvider } from '../../designSystem/';
-import { flamesSheet, flamesData } from '../../designSystem/assets/sprites/flames';
-import { Clock } from '../Clock';
+import { GameLoopContextProvider } from '../../designSystem/';
 import { GameBoard } from '../GameBoard';
+import { GameUI } from './gameScreen/GameUI';
 
 export const GameScreen = ({ gameBoardType, win, lose }) => {
-  const clock = useRef(null);
-
   return (
     <GameLoopContextProvider>
       <GameBoard type={gameBoardType} />
-      <Button
-        title='Win'
-        fontFamily='neuroBold'
-        onPress={() => win(clock.current.currentTime)}
-      />
-      <VSpacer height={20} />
-      <Button
-        title='Lose'
-        fontFamily='neuroBold'
-        onPress={() => lose(clock.current.currentTime)}
-      />
-      <VSpacer height={10} />
-      <Clock ref={clock} />
-      <SpriteSheet fps={30} data={flamesData} image={flamesSheet} />
+      <GameUI win={win} lose={lose} />
     </GameLoopContextProvider>
   );
 };
