+++ src/designSystem/index.ts
@@ -1,13 +1,13 @@
 export { AppContainer } from './layout/containers/AppContainer';
 export { ViewContainer } from './layout/containers/ViewContainer';
-export { ViewContainerAbsolute } from './layout/containers/ViewContainerAbsolute';
-export { ViewContainerFlex } from './layout/containers/ViewContainerFlex';
+export { ViewContainerAbsolute } from './layout/containers/viewContainer/ViewContainerAbsolute';
+export { ViewContainerFlex } from './layout/containers/viewContainer/ViewContainerFlex';
 export { GameWrapper } from './layout/containers/GameWrapper';
 export { Label } from './ui/Label';
 export { Button } from './ui/Button';
 export { VSpacer } from './layout/VSpacer';
 export { SpriteSheet } from './ui/SpriteSheet';
-export * from './context/gameLoop/GameLoopContext'
+export * from './context/gameLoop/GameLoopContext';
 export { GameLoopContextProvider } from './context/gameLoop/GameLoopContextProvider';
 export { GameAudioContextProvider } from './context/audio/GameAudioContextProvider';
 export { GameAudioContext } from './context/audio/GameAudioContext';
