+++ src/designSystem/layout/containers/gameWrapper/Header.tsx
@@ -1,7 +1,7 @@
 import { Button } from '../../../ui/Button';
 import { useContext } from 'react';
 import { GameAudioContext } from '../../../context/audio/GameAudioContext';
-import { ViewContainerFlex } from '../ViewContainerFlex';
+import { ViewContainerFlex } from '../viewContainer/ViewContainerFlex';
 
 export const Header = () => {
   const { musicMuted, muteMusic } = useContext(GameAudioContext);
