+++ src/designSystem/layout/containers/viewContainer/ViewContainerAbsolute.tsx
@@ -1,4 +1,4 @@
-import { ViewContainer } from './ViewContainer';
+import { ViewContainer } from '../ViewContainer';
 
 export const ViewContainerAbsolute = ({ children }) => {
   return <ViewContainer absolute>{children}</ViewContainer>;
