+++ src/designSystem/layout/containers/viewContainer/ViewContainerFlex.tsx
@@ -1,4 +1,4 @@
-import { ViewContainer } from './ViewContainer';
+import { ViewContainer } from '../ViewContainer';
 
 export const ViewContainerFlex = ({ children }) => {
   return <ViewContainer flex>{children}</ViewContainer>;
